cloud-run/application/app/requirements.txt
setuptools
importlib-metadata
numpy<2.0
dateparser==1.1.0
fastapi
google-cloud-firestore==2.13.1
google-cloud-bigquery[pandas]==3.11.0
google-cloud-logging==3.5.0
google-cloud-secret-manager==2.16.2
gunicorn==20.1.0
ib-insync==0.9.86
statsmodels==0.14.0
uvicorn
packaging
